requirements.txt
fastapi
uvicorn
openai
python-dotenv

# TO RUN THE APPLICATION
# uvicorn app.main:app --reload 
